SurfaceFlinger: OMAP extension to extend hwc api
Support an extension to the HWC callback mechanism, so
that the api can be freely extended in the future.
This change is intended to keep ABI compatibility for
the HWC as this api will only be used if HWC_EXTENDED_API
is set in the flags.
This requires a corresponding change in libhardware.

Change-Id: Ib916dd574cc82f10ea7443feef10214ecac9fea6
Signed-off-by: Tony Lofthouse <a0741364@ti.com>
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>
Adapted to N's SF/HWC1 from TI Lollipop changes
Added __unused attrs in hook_extension_cb()
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>