HWComposer: Pass layer stack id to HWC
Upon connection of external display SurfaceFlinger assigns it the
same layer stack as for the primary display to enable cloning. The
primary layer stack is then rendered for each display separately,
which doubles GPU load. In most cases cloning can be done more
efficiently in HWC, but in order to do it HWC should know which
displays are showing cloned composition. By querying layer stack
for each display HWC can decide if it should do the cloning or not.

Change-Id: Iecc16e8ff09d4e63b30d8897b1257071f94feb3e
Signed-off-by: Mykola Ostrovskyy <mykola@ti.com>
Adapted to N's SF/HWC1 from TI Lollipop changes
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>