egl: [TMP/HACK] Add back b/10194508 workaround.
This was removed because it did not pass CTS and was applied across the board
for all OMAP4 devices, but some still-shipping OMAP4 devices don't need it
due to using DDK 1.12 drivers. Those drivers are only for OMAP4470 though,
and other devices besides OMAPs depend on this workaround too, and we don't
care about CTS for them anyway.

This patch adds this previously-long-standing behavior (EGL_NATIVE_VISUAL_ID)
back under the opt-in BOARD_EGL_WORKAROUND_BUG_10194508 flag.

Change-Id: Ie611634952a35dc756f1de806c850235caf73544
Signed-off-by: D. Andrei Măceș <dmaces@nd.edu>